+++ modules/recon/allowed_methods.rb
@@ -63,11 +63,15 @@ class Arachni::Modules::AllowedMethods < Arachni::Module::Base
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>',
             version:     '0.1.4',
             targets:     %w(Generic),
+            references:  {
+                'Apache.org' => 'http://httpd.apache.org/docs/2.2/mod/core.html#limitexcept'
+            },
             issue:       {
                 name:            %q{Allowed HTTP methods},
                 description:     %q{The webserver claims that it supports the logged methods.},
                 tags:            %w(http methods options),
-                severity:        Severity::INFORMATIONAL
+                severity:        Severity::INFORMATIONAL,
+                remedy_guidance: %q{Configure your web server to disallow unnecessary HTTP method.}
             }
         }
     end