+++ modules/recon/backup_files.rb
@@ -61,15 +61,21 @@ class Arachni::Modules::BackupFiles < Arachni::Module::Base
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@gmail.com> ',
             version:     '0.2.2',
             targets:     %w(Generic),
+            references: {
+                'WebAppSec' => 'http://www.webappsec.org/projects/threat/classes/information_leakage.shtml'
+            },
             issue:       {
                 name:            %q{A backup file exists on the server.},
                 description:     %q{The server response indicates that a file matching
     the name of a common naming scheme for file backups can be publicly accessible.
     A developer has probably forgotten to remove this file after testing.
     This can lead to source code disclosure and privileged information leaks.},
-                tags: %w(path backup file discovery),
+                tags:            %w(path backup file discovery),
                 cew:             '530',
-                severity:        Severity::MEDIUM
+                severity:        Severity::MEDIUM,
+                remedy_guidance: %q{Do not keep alternative versions of files underneath the virtual web server root.
+                    When updating the site, delete or move the files to a directory outside the virtual root, edit them there,
+                    and move (or copy) the files back to the virtual root. Make sure that only the files that are actually in use reside under the virtual root.}
             }
 
         }