+++ modules/recon/common_directories.rb
@@ -50,11 +50,16 @@ class Arachni::Modules::CommonDirectories < Arachni::Module::Base
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@gmail.com> ',
             version:     '0.2.1',
             targets:     %w(Generic),
+            references: {
+                'CWE'   => 'http://cwe.mitre.org/data/definitions/538.html',
+                'OWASP' => 'https://www.owasp.org/index.php/Forced_browsing'
+            },
             issue:       {
-                name:     %q{A common directory exists on the server.},
-                tags:     %w(path directory common discovery),
-                cwe:      '538',
-                severity: Severity::MEDIUM
+                name:            %q{A common directory exists on the server.},
+                tags:            %w(path directory common discovery),
+                cwe:             '538',
+                severity:        Severity::MEDIUM,
+                remedy_guidance: %q{Do not expose file and directory information to the user.}
             }
 
         }