+++ modules/recon/http_put.rb
@@ -56,12 +56,16 @@ class Arachni::Modules::HTTP_PUT < Arachni::Module::Base
             author:      'Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>',
             version:     '0.1.4',
             targets:     %w(Generic),
+            references: {
+                'W3' => 'http://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html'
+            },
             issue:       {
-                name:        %q{HTTP PUT is enabled.},
-                description: %q{3rd parties can upload files to the web-server.},
-                tags:        %w(http methods put server),
-                cwe:         '650',
-                severity:    Severity::HIGH
+                name:            %q{HTTP PUT is enabled.},
+                description:     %q{3rd parties can upload files to the web-server.},
+                tags:            %w(http methods put server),
+                cwe:             '650',
+                severity:        Severity::HIGH,
+                remedy_guidance: %q{Disable the PUT method on the Web Server and/or disable write permissions to the web server directory.}
             }
         }
     end